Add github action workflow for windows testing (#434)
Make pkg/test.Exec work with windows. Remove the reference to bash
and call the command directly. Make sure the path arg to exec.Command
is a path, not just a command name, otherwise we'll trigger the usage
of exec.LookPath on windows, which expects to use the PATH in the
current process environment, and can't be overridden without an os.Setenv.

Make it so shell script system tests only build/run on linux amd mac
hosts.

Co-authored-by: Alfred Landrum <alfred@leakybucket.org>